After test on station
Had error in send data  not right data type

+++ Inc/uart_comm.h
@@ -18,7 +18,7 @@
 
  }_UART_BUF;
 
-ErrorStatus Data_transmite_UART_9B (uint16_t mass[], USART_TypeDef *USARTx);
+ErrorStatus Data_transmite_UART_9B (uint8_t mass[], USART_TypeDef *USARTx);
 void Parser_command ( _UART_BUF uart_receive_buffer, _SETTINGS_MODULE * module_settings, _ADC_PARAMETERS *adc_parametrs ,enum PWR_TIMx timer_numberr,  USART_TypeDef *USARTx);
